backend/apps/services/ClusteringService/requirements.txt
pytesseract>=0.3.10
pillow>=9.0.0
opencv-python>=4.5.5.64
numpy>=1.21.0
scikit-learn>=1.0.0
